+++ apps/nezon/src/types/index.ts
@@ -40,6 +40,7 @@ import * as Messaging from '../messaging/smart-message';
 import type {
   ManagedMessage as ManagedMessageType,
   DMHelper as DMHelperType,
+  ChannelHelper as ChannelHelperType,
   SmartMessageLike,
 } from '../messaging/smart-message';
 import * as ButtonModule from '../messaging/button-builder';
@@ -78,15 +79,21 @@ export namespace Nezon {
   export type ComponentParams = string[] | Record<string, string>;
   export type ComponentParam = string | undefined;
   export type Args = string[];
-  export type AutoContext = [ManagedMessageType, DMHelperType];
+  export type AutoContext = [
+    ManagedMessageType,
+    DMHelperType,
+    ChannelHelperType,
+  ];
   export type SmartMessageInput = SmartMessageLike;
   export const SmartMessage = Messaging.SmartMessage;
   export type ManagedMessage = ManagedMessageType;
   export type DMHelper = DMHelperType;
+  export type ChannelHelper = ChannelHelperType;
 
   export namespace AutoContextType {
     export type Message = ManagedMessageType;
     export type DM = DMHelperType;
+    export type Channel = ChannelHelperType;
   }
 
   export type Button = ButtonComponent;